+++ packages/Webkul/Admin/src/Routes/configuration-routes.php
@@ -12,10 +12,10 @@ Route::get('configuration/search', [ConfigurationController::class, 'search'])
     ->name('admin.configuration.search');
 
 Route::post('configuration/cache-management/execute', [CacheManagementController::class, 'execute'])
-    ->name('admin.configuration.cache-management.execute');
+    ->name('admin.configuration.cache_management.execute');
 
 Route::post('configuration/search-engines/{engine}/test-connection', [SearchEngineController::class, 'testConnection'])
-    ->name('admin.configuration.search-engines.test-connection');
+    ->name('admin.configuration.search_engines.test_connection');
 
 Route::controller(ConfigurationController::class)->prefix('configuration/{slug?}/{slug2?}')->group(function () {
     Route::get('', 'index')->name('admin.configuration.index');