+++ packages/Webkul/Shop/src/Resources/views/customers/account/eu-withdrawals/button.blade.php
@@ -15,14 +15,14 @@
     @if ($existingWithdrawal)
         <a
             class="{{ $variant ?? 'secondary-button border-zinc-200 px-5 py-3 font-normal max-md:hidden' }}"
-            href="{{ route('shop.customers.account.eu-withdrawal.show', $existingWithdrawal->uuid) }}"
+            href="{{ route('shop.customers.account.eu_withdrawal.show', $existingWithdrawal->uuid) }}"
         >
             @lang('shop::app.eu_withdrawal.button.view_existing')
         </a>
     @else
         <a
             class="{{ $variant ?? 'primary-button px-5 py-3 font-normal max-md:hidden' }}"
-            href="{{ route('shop.customers.account.eu-withdrawal.create', $order->id) }}"
+            href="{{ route('shop.customers.account.eu_withdrawal.create', $order->id) }}"
         >
             @lang('shop::app.eu_withdrawal.button.withdraw')
         </a>