+++ packages/Webkul/Shop/src/Resources/views/customers/account/rma/view.blade.php
@@ -132,8 +132,8 @@
                         <x-shop::form
                             enctype="multipart/form-data"
                             :action="$canCloseRma
-                                ? route('shop.customers.account.rma.update-status', $rma->id)
-                                : route('shop.customers.account.rma.re-open', $rma->id)"
+                                ? route('shop.customers.account.rma.update_status', $rma->id)
+                                : route('shop.customers.account.rma.re_open', $rma->id)"
                         >
                             @php $checkboxName = $canCloseRma ? 'close_rma' : 'reopen_rma'; @endphp
 
@@ -579,7 +579,7 @@
 
                 methods: {
                     getMessage() {
-                        this.$axios.get(`{{ route('shop.customers.account.rma.get-messages') }}`, {
+                        this.$axios.get(`{{ route('shop.customers.account.rma.get_messages') }}`, {
                             params: { id: this.rma.id, limit: this.limit }
                         })
                         .then(response => {
@@ -594,7 +594,7 @@
                         formData.set('message', sanitizedMessage);
                         this.isChatSend = false;
 
-                        this.$axios.post("{{ route('shop.customers.account.rma.send-message') }}", formData)
+                        this.$axios.post("{{ route('shop.customers.account.rma.send_message') }}", formData)
                             .then((response) => {
                                 const attachmentPreview = document.getElementById('attachmentPreview');
                                 attachmentPreview.innerHTML = '';