+++ packages/Webkul/SocialLogin/src/Http/routes.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Route;
 use Webkul\SocialLogin\Http\Controllers\LoginController;
 
 Route::controller(LoginController::class)->middleware(['web', 'shop'])->prefix('customer/social-login/{provider}')->group(function () {
-    Route::get('', 'redirectToProvider')->name('customer.social-login.index');
+    Route::get('', 'redirectToProvider')->name('customer.social_login.index');
 
-    Route::get('callback', 'handleProviderCallback')->name('customer.social-login.callback');
+    Route::get('callback', 'handleProviderCallback')->name('customer.social_login.callback');
 });